--- Assets/Scripts/GameManagerExample.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff450bf078e26b04ba2943153ef1c533
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
